Add score details to /projects/{id}
If the `score-detail=true` query parameter is passed along with
`full=true`, then a `detail` object is added to each fact that contains
information about how the score was calculated and options that
influence its value. This can be utilitized to provide a nice "on hover"
information panel for each fact value.